BehavioralEconomicR {BehavioralEconomicR}	R Documentation

Paquete de R disenado para trabajar de forma sencilla data de investigaciones de Economia Conductual. Creado por BEST

Description

Paquete de R disenado para trabajar de forma sencilla data de investigaciones de Economia Conductual. Creado por BEST

[Package BehavioralEconomicR version 0.1.1 Index]