ie_data {BehavioralEconomicR}	R Documentation

Paquete de R disenado para trabajar de forma sencilla data de investigaciones de Economia Conductual.

Description

Paquete de R disenado para trabajar de forma sencilla data de investigaciones de Economia Conductual.

Usage

data(ie_data)

Format

Base de datos del libro Irrational Exuberance de Roberth Shiller, premio Nobel de Economia por su aportes en finanzas conductuales. Esta data es actualizada mensualmente por el propio autor y la Universidad de Yale.

[Package BehavioralEconomicR version 0.1.1 Index]